requirements.txt
pymupdf>=1.23.0
pymongo>=4.6.0
httpx>=0.25.0
pandas>=2.0.0
python-dotenv>=1.0.0
